requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
flask==3.0.0
waitress==2.1.2
